feat/#117 : 스터디원 추천 시스템 구현

+++ Android/app/src/main/java/com/example/firstproject/data/model/dto/response/Top3RecommendedUsersDtoItem.kt
@@ -6,7 +6,7 @@ data class Top3RecommendedUsersDtoItem(
     val image: String,
     val meetingDays: List<String>,
     val nickname: String,
-    val similarity: Int,
+    val similarity: Double,
     val term: String,
     val topics: List<String>,
     val userId: String
